Fix issue with abstract classes
The ``getEntityClasses()`` method of ``SettingsCommand`` also return the class ``Gedmo\Tree\Entity\MappedSuperclass\AbstractClosure``.
This is because I use the Gedmo\Tree extension. This class is abstract and so, cannot be instanciated.

The ``discoverEntity`()` does not take care of abstract classes anymore, which can be usefull for other weird cases.