Allow skipping non Algolia entities in commands
When invoking the clear or reindex command for all known entities
the command will fail if at least one entity is not Algolia mapped.
This is incosistent regarding the settings command and makes it
difficult in automation processes.

By adding a new optional command line settings this kind of error
can be skipped. By default the command behaviour is not changing.